django_api/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
pandas==2.2.0
python-dotenv==1.0.1
beautifulsoup4==4.12.2
openpyxl==3.1.2
requests==2.31.0
selenium==4.16.0
webdriver-manager==4.0.1
pywhatkit==5.4
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.6
jinja2==3.1.2
python-telegram-bot==20.8
unidecode==1.3.8
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
schedule==1.2.1
pyOpenSSL
django-extensions==3.2.3
Werkzeug==3.0.1 
djangorestframework-simplejwt 
Pillow